Refactor file structure for components
