+++ src/components/app/app.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Main from "./main";
+import Main from "../main/main.jsx";
 
 const App = (props) => {
   // eslint-disable-next-line react/prop-types
