+++ src/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/app";
+import App from "./components/app/app.jsx";
 
-const OFFER_PLACES_COUNT = 313;
+const OFFER_PLACES_COUNT = 312;
 
 ReactDOM.render(
     <App
